+++ sample/cxx/v3/kind.pb.cc
@@ -974,94 +974,95 @@ const char descriptor_table_protodef_kind_2eproto[] ABSL_ATTRIBUTE_SECTION_VARIA
     "igin_timepoint_arr\030\032 \003(\t\022O\n\024convert_dura"
     "tion_arr\030\033 \003(\0132\031.google.protobuf.Duratio"
     "nB\026\302\?\023origin_duration_arr\022\033\n\023origin_dura"
-    "tion_arr\030\034 \003(\t\"\327\001\n\020role_upgrade_cfg\022\037\n\002I"
+    "tion_arr\030\034 \003(\t\"\350\001\n\020role_upgrade_cfg\022\037\n\002I"
     "d\030\001 \001(\rB\023\312>\014custom_rule3\350D\350\007\022\022\n\005Level\030\002 "
     "\001(\rB\003\350D\001\0226\n\010CostType\030\003 \001(\005B$\312>\014custom_ru"
-    "le1\232\?\022Refer to cost_type\022\021\n\tCostValue\030\004 "
-    "\001(\003\022\020\n\010ScoreAdd\030\005 \001(\005:1\312>%Test role_upgr"
-    "ade_cfg with multi keys\352D\006helper\"\353\004\n\narr"
-    "_in_arr\022-\n\004name\030\001 \001(\tB\037\232\?\034This is a test"
-    " name in array\022;\n\007int_arr\030\002 \003(\005B*\320\?\001\332\? g"
-    "ame_const_config.EN_GCC_UNKNOWN\342\?\0013\022\017\n\007s"
-    "tr_arr\030\003 \003(\t\022#\n\016test_info_role\030\013 \001(\0132\t.r"
-    "ole_cfgH\000\0222\n\025test_role_upgrade_cfg\030\014 \001(\013"
-    "2\021.role_upgrade_cfgH\000\032\373\001\n\023test_nested_me"
-    "ssage\0222\n\035test_nested_message_info_role\030\013"
-    " \001(\0132\t.role_cfgH\000\022A\n$test_nested_message"
-    "_role_upgrade_cfg\030\014 \001(\0132\021.role_upgrade_c"
-    "fgH\000\"`\n\020test_nested_enum\022%\n!EN_TEST_NEST"
-    "ED_MESSAGE_ENUM_VAL_1\020\000\022%\n!EN_TEST_NESTE"
-    "D_MESSAGE_ENUM_VAL_2\020{B\013\n\ttest_onof\"P\n\020t"
-    "est_nested_enum\022\035\n\031EN_TEST_NESTED_ENUM_V"
-    "AL_1\020\000\022\035\n\031EN_TEST_NESTED_ENUM_VAL_2\020{:*\352"
-    "D\006helper\360D\001\202E\033Engine/CompositeDataTable."
-    "hB\013\n\ttest_onof\"}\n\021test_msg_verifier\022\022\n\tt"
-    "est_id_1\030\221N \001(\r\0224\n\ttest_id_2\030\222N \001(\rB \242\?\017"
-    "\346\265\213\350\257\225ID\345\210\253\345\220\2152\362\?\013server_only\022\027\n\ttest_id"
-    "_3\030\223N \001(\rB\003\310\?\001:\005\242\?\002-^\"\232\004\n\016arr_in_arr_cfg"
-    "\022\037\n\002id\030\001 \001(\rB\023\232\?\rThis is a Key\350D\001\022(\n\003arr"
-    "\030\002 \003(\0132\013.arr_in_arrB\016\362\?\013server_only\0222\n\022t"
-    "est_plain_int_arr\030\003 \003(\005B\026\332\?\0011\342\?\0013\362\?\013serv"
-    "er_only\022\'\n\023test_plain_enum_arr\030\004 \003(\0162\n.c"
-    "ost_type\0220\n\016test_plain_msg\030\005 \001(\0132\022.test_"
-    "msg_verifierB\004\262\?\001&\022.\n\022test_plain_msg_arr"
-    "\030\006 \003(\0132\022.test_msg_verifier\0223\n\013test_map_i"
-    "s\030\007 \003(\0132\036.arr_in_arr_cfg.TestMapIsEntry\022"
-    "9\n\013test_map_sm\030\010 \003(\0132\036.arr_in_arr_cfg.Te"
-    "stMapSmEntryB\004\262\?\001|\0320\n\016TestMapIsEntry\022\013\n\003"
-    "key\030\001 \001(\005\022\r\n\005value\030\002 \001(\t:\0028\001\032;\n\016TestMapS"
-    "mEntry\022\013\n\003key\030\001 \001(\t\022\030\n\005value\030\002 \001(\0132\t.dep"
-    "2_cfg:\0028\001:\037\312>\023Test arr_in_arr_cfg\352D\006help"
-    "er\"\243\001\n\021event_reward_item\022\017\n\007item_id\030\001 \001("
-    "\r\022\022\n\nitem_count\030\002 \001(\r\022&\n\013nested_note\030\013 \001"
-    "(\tB\017\242\?\014\346\217\217\350\277\260\346\226\207\346\234\254H\000\0227\n\020nested_enum_typ"
-    "e\030\014 \001(\0162\n.cost_typeB\017\242\?\014\350\264\247\345\270\201\347\261\273\345\236\213H\000B\010"
-    "\n\006nested\"\241\001\n\017event_rule_item\022\017\n\007rule_id\030"
-    "\001 \001(\r\022\022\n\nrule_param\030\002 \001(\r\022&\n\013nested_note"
-    "\030\013 \001(\tB\017\242\?\014\346\217\217\350\277\260\346\226\207\346\234\254H\000\0227\n\020nested_enum"
-    "_type\030\014 \001(\0162\n.cost_typeB\017\242\?\014\350\264\247\345\270\201\347\261\273\345\236\213"
-    "H\000B\010\n\006nested\"\302\004\n\tevent_cfg\022.\n\002id\030\001 \001(\rB\""
-    "\312>\014custom_rule4\322>\nunique_tag\330>\001\350Dd\022$\n\007pr"
-    "ocess\030\002 \001(\rB\023\322>\nunique_tag\330>\001\350D\001\022\036\n\004rule"
-    "\030\006 \001(\0132\020.event_rule_item\0227\n\rspecify_fiel"
-    "d\030\007 \001(\0132\020.event_rule_itemB\016\362\?\013server_onl"
-    "y\0223\n\004item\030\013 \001(\0132\022.event_reward_itemB\017\242\?\014"
-    "\345\245\226\345\212\261\351\201\223\345\205\267H\000\022#\n\010user_exp\030\014 \001(\003B\017\242\?\014\345\245\226"
-    "\345\212\261\347\273\217\351\252\214H\000\022\037\n\004note\030\r \001(\tB\017\242\?\014\346\217\217\350\277\260\346\226\207\346"
-    "\234\254H\000\0220\n\tenum_type\030\016 \001(\0162\n.cost_typeB\017\242\?\014"
-    "\350\264\247\345\270\201\347\261\273\345\236\213H\000\022%\n\nuser_level\0303 \001(\rB\017\242\?\014\347"
-    "\216\251\345\256\266\347\255\211\347\272\247H\001\022(\n\010test_arr\030[ \003(\0132\020.event_"
-    "rule_itemB\004\262\?\001;\022.\n\016test_empty_arr\030\\ \003(\0132"
-    "\020.event_rule_itemB\004\262\?\001;:,\312> Test event_c"
-    "fg with oneof fields\352D\006helperB\033\n\006reward\022"
-    "\021\330>\001\352>\013server_onlyB\r\n\013unlock_type\"\266\001\n\034ke"
-    "ep_or_strip_empty_list_cfg\022\037\n\002id\030\001 \001(\rB\023"
-    "\232\?\rThis is a Key\350D\001\022\034\n\tarray_msg\030\002 \003(\0132\t"
-    ".dep2_cfg\022\"\n\017array_plain_msg\030\003 \003(\0132\t.dep"
-    "2_cfg\022\023\n\013array_int32\030\004 \003(\005\022\023\n\013array_int6"
-    "4\030\005 \003(\003:\t\352D\006helper\"4\n\ncombine_id\022\016\n\006pref"
-    "ix\030\001 \001(\r\022\026\n\002id\030\002 \001(\rB\n\312>\0070-99999\"\333\003\n\021pro"
-    "cess_by_script\022\n\n\002id\030\001 \001(\r\022<\n\016map_u32_st"
-    "ring\030\002 \003(\0132$.process_by_script.MapU32Str"
-    "ingEntry\022\033\n\023human_readable_date\030\003 \001(\t\022(\n"
-    "\004date\030\004 \001(\0132\032.google.protobuf.Timestamp\022"
-    "6\n\013test_map_is\030\007 \003(\0132!.process_by_script"
-    ".TestMapIsEntry\022<\n\013test_map_sm\030\010 \003(\0132!.p"
-    "rocess_by_script.TestMapSmEntryB\004\262\?\001|\022\033\n"
-    "\005id_id\030\351\007 \001(\0132\013.combine_id\0323\n\021MapU32Stri"
-    "ngEntry\022\013\n\003key\030\001 \001(\r\022\r\n\005value\030\002 \001(\t:\0028\001\032"
-    "0\n\016TestMapIsEntry\022\013\n\003key\030\001 \001(\005\022\r\n\005value\030"
-    "\002 \001(\t:\0028\001\032;\n\016TestMapSmEntry\022\013\n\003key\030\001 \001(\t"
-    "\022\030\n\005value\030\002 \001(\0132\t.dep2_cfg:\0028\001\"\222\001\n\017large"
-    "_file_test\022\n\n\002id\030\001 \001(\r\022\020\n\010i32value\030\002 \001(\005"
-    "\022\020\n\010u64value\030\003 \001(\004\022\020\n\010f64value\030\004 \001(\001\022\020\n\010"
-    "f32value\030\005 \001(\002\022\017\n\007szvalue\030\006 \001(\t\022\032\n\005id_id"
-    "\030\007 \001(\0132\013.combine_id\"N\n\023inner_alias_messa"
-    "ge\022\026\n\005iType\030\001 \001(\005B\007\242\?\004Type\022\037\n\tparamList\030"
-    "\002 \003(\003B\014\242\?\005Param\332\?\0012\"|\n\023outer_alias_messa"
-    "ge\022\022\n\003iId\030\001 \001(\005B\005\242\?\002ID\0222\n\006inners\030\002 \003(\0132\024"
-    ".inner_alias_messageB\014\242\?\tCondition\022\035\n\010ne"
-    "st_int\030\003 \003(\005B\013\242\?\004Nest\332\?\0012b\006proto3"
+    "le1\232\?\022Refer to cost_type\022\"\n\tCostValue\030\004 "
+    "\001(\003B\017\312>\014custom_rule5\022\020\n\010ScoreAdd\030\005 \001(\005:1"
+    "\312>%Test role_upgrade_cfg with multi keys"
+    "\352D\006helper\"\353\004\n\narr_in_arr\022-\n\004name\030\001 \001(\tB\037"
+    "\232\?\034This is a test name in array\022;\n\007int_a"
+    "rr\030\002 \003(\005B*\320\?\001\332\? game_const_config.EN_GCC"
+    "_UNKNOWN\342\?\0013\022\017\n\007str_arr\030\003 \003(\t\022#\n\016test_in"
+    "fo_role\030\013 \001(\0132\t.role_cfgH\000\0222\n\025test_role_"
+    "upgrade_cfg\030\014 \001(\0132\021.role_upgrade_cfgH\000\032\373"
+    "\001\n\023test_nested_message\0222\n\035test_nested_me"
+    "ssage_info_role\030\013 \001(\0132\t.role_cfgH\000\022A\n$te"
+    "st_nested_message_role_upgrade_cfg\030\014 \001(\013"
+    "2\021.role_upgrade_cfgH\000\"`\n\020test_nested_enu"
+    "m\022%\n!EN_TEST_NESTED_MESSAGE_ENUM_VAL_1\020\000"
+    "\022%\n!EN_TEST_NESTED_MESSAGE_ENUM_VAL_2\020{B"
+    "\013\n\ttest_onof\"P\n\020test_nested_enum\022\035\n\031EN_T"
+    "EST_NESTED_ENUM_VAL_1\020\000\022\035\n\031EN_TEST_NESTE"
+    "D_ENUM_VAL_2\020{:*\352D\006helper\360D\001\202E\033Engine/Co"
+    "mpositeDataTable.hB\013\n\ttest_onof\"}\n\021test_"
+    "msg_verifier\022\022\n\ttest_id_1\030\221N \001(\r\0224\n\ttest"
+    "_id_2\030\222N \001(\rB \242\?\017\346\265\213\350\257\225ID\345\210\253\345\220\2152\362\?\013serve"
+    "r_only\022\027\n\ttest_id_3\030\223N \001(\rB\003\310\?\001:\005\242\?\002-^\"\232"
+    "\004\n\016arr_in_arr_cfg\022\037\n\002id\030\001 \001(\rB\023\232\?\rThis i"
+    "s a Key\350D\001\022(\n\003arr\030\002 \003(\0132\013.arr_in_arrB\016\362\?"
+    "\013server_only\0222\n\022test_plain_int_arr\030\003 \003(\005"
+    "B\026\332\?\0011\342\?\0013\362\?\013server_only\022\'\n\023test_plain_e"
+    "num_arr\030\004 \003(\0162\n.cost_type\0220\n\016test_plain_"
+    "msg\030\005 \001(\0132\022.test_msg_verifierB\004\262\?\001&\022.\n\022t"
+    "est_plain_msg_arr\030\006 \003(\0132\022.test_msg_verif"
+    "ier\0223\n\013test_map_is\030\007 \003(\0132\036.arr_in_arr_cf"
+    "g.TestMapIsEntry\0229\n\013test_map_sm\030\010 \003(\0132\036."
+    "arr_in_arr_cfg.TestMapSmEntryB\004\262\?\001|\0320\n\016T"
+    "estMapIsEntry\022\013\n\003key\030\001 \001(\005\022\r\n\005value\030\002 \001("
+    "\t:\0028\001\032;\n\016TestMapSmEntry\022\013\n\003key\030\001 \001(\t\022\030\n\005"
+    "value\030\002 \001(\0132\t.dep2_cfg:\0028\001:\037\312>\023Test arr_"
+    "in_arr_cfg\352D\006helper\"\243\001\n\021event_reward_ite"
+    "m\022\017\n\007item_id\030\001 \001(\r\022\022\n\nitem_count\030\002 \001(\r\022&"
+    "\n\013nested_note\030\013 \001(\tB\017\242\?\014\346\217\217\350\277\260\346\226\207\346\234\254H\000\0227"
+    "\n\020nested_enum_type\030\014 \001(\0162\n.cost_typeB\017\242\?"
+    "\014\350\264\247\345\270\201\347\261\273\345\236\213H\000B\010\n\006nested\"\241\001\n\017event_rule"
+    "_item\022\017\n\007rule_id\030\001 \001(\r\022\022\n\nrule_param\030\002 \001"
+    "(\r\022&\n\013nested_note\030\013 \001(\tB\017\242\?\014\346\217\217\350\277\260\346\226\207\346\234\254"
+    "H\000\0227\n\020nested_enum_type\030\014 \001(\0162\n.cost_type"
+    "B\017\242\?\014\350\264\247\345\270\201\347\261\273\345\236\213H\000B\010\n\006nested\"\302\004\n\tevent_"
+    "cfg\022.\n\002id\030\001 \001(\rB\"\312>\014custom_rule4\322>\nuniqu"
+    "e_tag\330>\001\350Dd\022$\n\007process\030\002 \001(\rB\023\322>\nunique_"
+    "tag\330>\001\350D\001\022\036\n\004rule\030\006 \001(\0132\020.event_rule_ite"
+    "m\0227\n\rspecify_field\030\007 \001(\0132\020.event_rule_it"
+    "emB\016\362\?\013server_only\0223\n\004item\030\013 \001(\0132\022.event"
+    "_reward_itemB\017\242\?\014\345\245\226\345\212\261\351\201\223\345\205\267H\000\022#\n\010user_"
+    "exp\030\014 \001(\003B\017\242\?\014\345\245\226\345\212\261\347\273\217\351\252\214H\000\022\037\n\004note\030\r \001"
+    "(\tB\017\242\?\014\346\217\217\350\277\260\346\226\207\346\234\254H\000\0220\n\tenum_type\030\016 \001(\016"
+    "2\n.cost_typeB\017\242\?\014\350\264\247\345\270\201\347\261\273\345\236\213H\000\022%\n\nuser_"
+    "level\0303 \001(\rB\017\242\?\014\347\216\251\345\256\266\347\255\211\347\272\247H\001\022(\n\010test_a"
+    "rr\030[ \003(\0132\020.event_rule_itemB\004\262\?\001;\022.\n\016test"
+    "_empty_arr\030\\ \003(\0132\020.event_rule_itemB\004\262\?\001;"
+    ":,\312> Test event_cfg with oneof fields\352D\006"
+    "helperB\033\n\006reward\022\021\330>\001\352>\013server_onlyB\r\n\013u"
+    "nlock_type\"\266\001\n\034keep_or_strip_empty_list_"
+    "cfg\022\037\n\002id\030\001 \001(\rB\023\232\?\rThis is a Key\350D\001\022\034\n\t"
+    "array_msg\030\002 \003(\0132\t.dep2_cfg\022\"\n\017array_plai"
+    "n_msg\030\003 \003(\0132\t.dep2_cfg\022\023\n\013array_int32\030\004 "
+    "\003(\005\022\023\n\013array_int64\030\005 \003(\003:\t\352D\006helper\"4\n\nc"
+    "ombine_id\022\016\n\006prefix\030\001 \001(\r\022\026\n\002id\030\002 \001(\rB\n\312"
+    ">\0070-99999\"\333\003\n\021process_by_script\022\n\n\002id\030\001 "
+    "\001(\r\022<\n\016map_u32_string\030\002 \003(\0132$.process_by"
+    "_script.MapU32StringEntry\022\033\n\023human_reada"
+    "ble_date\030\003 \001(\t\022(\n\004date\030\004 \001(\0132\032.google.pr"
+    "otobuf.Timestamp\0226\n\013test_map_is\030\007 \003(\0132!."
+    "process_by_script.TestMapIsEntry\022<\n\013test"
+    "_map_sm\030\010 \003(\0132!.process_by_script.TestMa"
+    "pSmEntryB\004\262\?\001|\022\033\n\005id_id\030\351\007 \001(\0132\013.combine"
+    "_id\0323\n\021MapU32StringEntry\022\013\n\003key\030\001 \001(\r\022\r\n"
+    "\005value\030\002 \001(\t:\0028\001\0320\n\016TestMapIsEntry\022\013\n\003ke"
+    "y\030\001 \001(\005\022\r\n\005value\030\002 \001(\t:\0028\001\032;\n\016TestMapSmE"
+    "ntry\022\013\n\003key\030\001 \001(\t\022\030\n\005value\030\002 \001(\0132\t.dep2_"
+    "cfg:\0028\001\"\222\001\n\017large_file_test\022\n\n\002id\030\001 \001(\r\022"
+    "\020\n\010i32value\030\002 \001(\005\022\020\n\010u64value\030\003 \001(\004\022\020\n\010f"
+    "64value\030\004 \001(\001\022\020\n\010f32value\030\005 \001(\002\022\017\n\007szval"
+    "ue\030\006 \001(\t\022\032\n\005id_id\030\007 \001(\0132\013.combine_id\"N\n\023"
+    "inner_alias_message\022\026\n\005iType\030\001 \001(\005B\007\242\?\004T"
+    "ype\022\037\n\tparamList\030\002 \003(\003B\014\242\?\005Param\332\?\0012\"|\n\023"
+    "outer_alias_message\022\022\n\003iId\030\001 \001(\005B\005\242\?\002ID\022"
+    "2\n\006inners\030\002 \003(\0132\024.inner_alias_messageB\014\242"
+    "\?\tCondition\022\035\n\010nest_int\030\003 \003(\005B\013\242\?\004Nest\332\?"
+    "\0012b\006proto3"
 };
 static const ::_pbi::DescriptorTable* const descriptor_table_kind_2eproto_deps[6] =
     {
@@ -1076,7 +1077,7 @@ static ::absl::once_flag descriptor_table_kind_2eproto_once;
 PROTOBUF_CONSTINIT const ::_pbi::DescriptorTable descriptor_table_kind_2eproto = {
     false,
     false,
-    4313,
+    4330,
     descriptor_table_protodef_kind_2eproto,
     "kind.proto",
     &descriptor_table_kind_2eproto_once,
@@ -2002,7 +2003,7 @@ const ::_pbi::TcParseTable<3, 5, 0, 0, 2> role_upgrade_cfg::_table_ = {
     // int32 CostType = 3 [(.org.xresloader.validator) = "custom_rule1", (.org.xresloader.field_description) = "Refer to cost_type"];
     {::_pbi::TcParser::SingularVarintNoZag1<::uint32_t, offsetof(role_upgrade_cfg, _impl_.costtype_), 63>(),
      {24, 63, 0, PROTOBUF_FIELD_OFFSET(role_upgrade_cfg, _impl_.costtype_)}},
-    // int64 CostValue = 4;
+    // int64 CostValue = 4 [(.org.xresloader.validator) = "custom_rule5"];
     {::_pbi::TcParser::SingularVarintNoZag1<::uint64_t, offsetof(role_upgrade_cfg, _impl_.costvalue_), 63>(),
      {32, 63, 0, PROTOBUF_FIELD_OFFSET(role_upgrade_cfg, _impl_.costvalue_)}},
     // int32 ScoreAdd = 5;
@@ -2022,7 +2023,7 @@ const ::_pbi::TcParseTable<3, 5, 0, 0, 2> role_upgrade_cfg::_table_ = {
     // int32 CostType = 3 [(.org.xresloader.validator) = "custom_rule1", (.org.xresloader.field_description) = "Refer to cost_type"];
     {PROTOBUF_FIELD_OFFSET(role_upgrade_cfg, _impl_.costtype_), 0, 0,
     (0 | ::_fl::kFcSingular | ::_fl::kInt32)},
-    // int64 CostValue = 4;
+    // int64 CostValue = 4 [(.org.xresloader.validator) = "custom_rule5"];
     {PROTOBUF_FIELD_OFFSET(role_upgrade_cfg, _impl_.costvalue_), 0, 0,
     (0 | ::_fl::kFcSingular | ::_fl::kInt64)},
     // int32 ScoreAdd = 5;
@@ -2083,7 +2084,7 @@ PROTOBUF_NOINLINE void role_upgrade_cfg::Clear() {
                     stream, this_._internal_costtype(), target);
           }
 
-          // int64 CostValue = 4;
+          // int64 CostValue = 4 [(.org.xresloader.validator) = "custom_rule5"];
           if (this_._internal_costvalue() != 0) {
             target = ::google::protobuf::internal::WireFormatLite::
                 WriteInt64ToArrayWithField<4>(
@@ -2132,7 +2133,7 @@ PROTOBUF_NOINLINE void role_upgrade_cfg::Clear() {
               total_size += ::_pbi::WireFormatLite::UInt32SizePlusOne(
                   this_._internal_level());
             }
-            // int64 CostValue = 4;
+            // int64 CostValue = 4 [(.org.xresloader.validator) = "custom_rule5"];
             if (this_._internal_costvalue() != 0) {
               total_size += ::_pbi::WireFormatLite::Int64SizePlusOne(
                   this_._internal_costvalue());