+++ sample/proto_v3/kind_option.js
@@ -486,6 +486,14 @@ exports.files = [
             },
             type_name : "INT32"
           },
+          CostValue : {
+            name : "CostValue",
+            number : 4,
+            options : {
+              validator : "custom_rule5"
+            },
+            type_name : "INT64"
+          },
           Id : {
             name : "Id",
             number : 1,